Problem: sit-git-hook is insufficient
While [sit-git-hook](https://github.com/sit-fyi/sit-git-hook)
has enabled exchange of sit repository information, it is
still not a full solution for setting up an exchange point,
it requires a further setup (http or ssh endpoint; or an email
processing endpoint), as it is just a hook.

As a matter of fact, even current setups, being hosted
on gitolite, suffer from being served over SSH as this
requires a private key (hence, trickier git setup) and
that key needs secure permissions, which is something Git
can't preserve.

Solution: develop sit-inbox to automate this whole setup
in one box (specifically, a Docker container, but can be
extended to other things later)

Its initial focus on decentralized workflows (like email)
with the intent to extend it to other workflows shortly.

The idea is that this "box" can be run on a maintainer's
machine where it'll be polling its inboxes and forwarding
valid additions into repositories. It can also run on a
server for either the email workflow or (future) Git endpoint
serving workflow.