Problem: repositories outside of .sit
In a variety of cases, we might have multiple
repositories within one directory, or have
them in non-standard locations.

For example, a repository can be in the root.

Alternatively, there might be a repository for
defined per sub-directory.

Solution: introduce repository.*.paths option
that defaults to `.sit` but allows specifying
multiple paths, and allows for wildcards.